+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v3.16.0
 ## 🚀 Feature
+- Added states to `AndesDropdownForm` | Author: [@MKiperszmid](https://github.com/MKiperszmid)
 - Added GetItems to `AndesDropdownForm` and `AndesDropdownStandalone` and Default Item to `AndesDropdownForm` | Author: [@MKiperszmid](https://github.com/MKiperszmid)
 - Added a SizeStyle to the Tooltip to show it as default or full screen.
 - Added Indeterminate support to `AndesTextfield` | Author: [@MKiperszmid](https://github.com/MKiperszmid)